Split long code files into several small ones
