+++ lib/features/home/service/components/appointment_card.dart
@@ -21,23 +21,20 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <https://www.gnu.org/licenses/>.
 ///
-/// Authors: Zheyuan Xu
+/// Authors: Zheyuan Xu, Tony Chen
 
 library;
 
 import 'package:flutter/material.dart';
 
-import 'package:audioplayers/audioplayers.dart';
-import 'package:intl/intl.dart';
 import 'package:markdown_tooltip/markdown_tooltip.dart';
 
 import 'package:healthpod/features/diary/models/appointment.dart';
 import 'package:healthpod/features/diary/service.dart';
+import 'package:healthpod/features/home/service/components/dialogs/appointment_management_dialog.dart';
+import 'package:healthpod/features/home/service/components/widgets/appointment_info_display.dart';
+import 'package:healthpod/features/home/service/components/widgets/transport_info_widget.dart';
 import 'package:healthpod/theme/card_style.dart';
-
-// Global flag to track if transport audio is currently playing.
-
-bool transportAudioIn = false;
 
 /// A widget that displays both next appointment details and appointment summary.
 ///
@@ -52,14 +49,6 @@ class AppointmentCard extends StatefulWidget {
 }
 
 class _AppointmentCardState extends State<AppointmentCard> {
-  // Flag indicating whether audio is currently playing.
-
-  bool _isPlaying = false;
-
-  // Audio player instance for handling transport eligibility audio.
-
-  final AudioPlayer _audioPlayer = AudioPlayer();
-
   // Card title displayed at the top of the component.
 
   String title = 'Medical Appointments';
@@ -100,49 +89,10 @@ class _AppointmentCardState extends State<AppointmentCard> {
 
   bool _isLoading = true;
 
-  /// Toggles the audio playback state.
-  ///
-  /// Stops playback if currently playing, or starts playback if stopped.
-  /// Ensures only one audio instance plays at a time.
-
-  Future<void> _toggleAudio() async {
-    if (_isPlaying) {
-      await _audioPlayer.stop();
-
-      setState(() {
-        _isPlaying = false;
-        transportAudioIn = false;
-      });
-    } else {
-      if (!transportAudioIn) {
-        await _audioPlayer.play(AssetSource('audio/transport_eligibility.mp3'));
-
-        setState(() {
-          _isPlaying = !_isPlaying;
-          transportAudioIn = true;
-        });
-      }
-    }
-  }
-
-  /// Handles the completion of audio playback.
-  ///
-  /// Resets the playing state and global audio flag when playback finishes.
-
-  void _onAudioComplete() {
-    setState(() {
-      _isPlaying = false;
-      transportAudioIn = false;
-    });
-  }
-
   @override
   void initState() {
     super.initState();
     _loadAppointments();
-    _audioPlayer.onPlayerComplete.listen((event) {
-      _onAudioComplete();
-    });
   }
 
   Future<void> _loadAppointments() async {
@@ -168,287 +118,27 @@ class _AppointmentCardState extends State<AppointmentCard> {
 
   @override
   void dispose() {
-    _audioPlayer.dispose();
-    transportAudioIn = false;
     super.dispose();
   }
 
   /// Opens a dialog to manage all appointments.
-  ///
-  /// Displays a list of current appointments with options to add, delete,
-  /// import, or export appointments.
 
   void _manageAppointments() {
     showDialog(
       context: context,
       builder: (BuildContext context) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            return AlertDialog(
-              title: const Text('Manage Appointments'),
-              content: SizedBox(
-                width: double.maxFinite,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    const Text(
-                      'Current Appointments',
-                      style: TextStyle(fontWeight: FontWeight.bold),
-                    ),
-                    const SizedBox(height: 8),
-                    Flexible(
-                      child: ListView.builder(
-                        shrinkWrap: true,
-                        itemCount: appointments.length,
-                        itemBuilder: (context, index) {
-                          final appointment = appointments[index];
-                          return Card(
-                            margin: const EdgeInsets.only(bottom: 8),
-                            child: ListTile(
-                              title: Text(appointment.title),
-                              subtitle: Text(
-                                '${DateFormat('d MMM yyyy').format(appointment.date)} at ${DateFormat('h:mm a').format(appointment.date)}\n${appointment.description}',
-                              ),
-                              trailing: IconButton(
-                                icon: const Icon(Icons.delete),
-                                onPressed: () async {
-                                  final success =
-                                      await DiaryService.deleteAppointment(
-                                    context,
-                                    appointment,
-                                  );
-                                  if (success) {
-                                    setState(() {
-                                      appointments.removeAt(index);
-                                    });
-                                  }
-                                },
-                              ),
-                              isThreeLine: true,
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Align(
-                      alignment: Alignment.center,
-                      child: ElevatedButton.icon(
-                        onPressed: () {
-                          // Show dialog to add new appointment.
-
-                          _showAddAppointmentDialog(context, setState);
-                        },
-                        icon: const Icon(Icons.add),
-                        label: const Text('Add Appointment'),
-                      ),
-                    ),
-                    const SizedBox(height: 16),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                      children: [
-                        // Import button.
-
-                        OutlinedButton.icon(
-                          onPressed: () {
-                            // TODO: Implement import functionality.
-
-                            Navigator.pop(context);
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(
-                                content: Text('Import feature coming soon'),
-                              ),
-                            );
-                          },
-                          icon: const Icon(Icons.upload_file),
-                          label: const Text('Import'),
-                        ),
-                        // Export button.
-
-                        OutlinedButton.icon(
-                          onPressed: () {
-                            // TODO: Implement export functionality.
-
-                            Navigator.pop(context);
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(
-                                content: Text('Export feature coming soon'),
-                              ),
-                            );
-                          },
-                          icon: const Icon(Icons.download),
-                          label: const Text('Export'),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-              actions: [
-                TextButton(
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                  child: const Text('Cancel'),
-                ),
-                ElevatedButton(
-                  onPressed: () {
-                    this.setState(() {
-                      // Update the next appointment if there are any appointments.
-
-                      if (appointments.isNotEmpty) {
-                        final nextAppointment = appointments.first;
-                        appointmentDate = nextAppointment.date;
-                        location = nextAppointment.description;
-                      }
-                    });
-                    Navigator.pop(context);
-                  },
-                  child: const Text('Save'),
-                ),
-              ],
-            );
+        return AppointmentManagementDialog(
+          appointments: appointments,
+          onAppointmentsUpdated: (updatedAppointments) {
+            setState(() {
+              appointments = updatedAppointments;
+              if (appointments.isNotEmpty) {
+                final nextAppointment = appointments.first;
+                appointmentDate = nextAppointment.date;
+                location = nextAppointment.description;
+              }
+            });
           },
-        );
-      },
-    );
-  }
-
-  /// Shows a dialog to add a new appointment.
-  ///
-  /// Provides form fields for appointment title, doctor name, location,
-  /// date, and time selection.
-
-  void _showAddAppointmentDialog(
-    BuildContext context,
-    void Function(void Function()) parentSetState,
-  ) {
-    final titleController = TextEditingController();
-    final descriptionController = TextEditingController();
-    DateTime selectedDate = DateTime.now();
-    TimeOfDay selectedTime = TimeOfDay.now();
-
-    showDialog(
-      context: context,
-      builder: (BuildContext dialogContext) {
-        return AlertDialog(
-          title: const Text('Add New Appointment'),
-          content: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                TextField(
-                  controller: titleController,
-                  decoration: const InputDecoration(
-                    labelText: 'Appointment Title',
-                    border: OutlineInputBorder(),
-                  ),
-                ),
-                const SizedBox(height: 16),
-                TextField(
-                  controller: descriptionController,
-                  decoration: const InputDecoration(
-                    labelText: 'Description',
-                    border: OutlineInputBorder(),
-                  ),
-                ),
-                const SizedBox(height: 16),
-                Row(
-                  children: [
-                    Expanded(
-                      child: OutlinedButton.icon(
-                        onPressed: () async {
-                          final DateTime? picked = await showDatePicker(
-                            context: dialogContext,
-                            initialDate: selectedDate,
-                            firstDate: DateTime.now(),
-                            lastDate: DateTime(2100),
-                          );
-                          if (picked != null) {
-                            selectedDate = DateTime(
-                              picked.year,
-                              picked.month,
-                              picked.day,
-                              selectedTime.hour,
-                              selectedTime.minute,
-                            );
-                          }
-                        },
-                        icon: const Icon(Icons.calendar_today),
-                        label: const Text('Set Date'),
-                      ),
-                    ),
-                    const SizedBox(width: 8),
-                    Expanded(
-                      child: OutlinedButton.icon(
-                        onPressed: () async {
-                          final TimeOfDay? picked = await showTimePicker(
-                            context: dialogContext,
-                            initialTime: selectedTime,
-                          );
-                          if (picked != null) {
-                            selectedTime = picked;
-                            selectedDate = DateTime(
-                              selectedDate.year,
-                              selectedDate.month,
-                              selectedDate.day,
-                              selectedTime.hour,
-                              selectedTime.minute,
-                            );
-                          }
-                        },
-                        icon: const Icon(Icons.access_time),
-                        label: const Text('Set Time'),
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.pop(dialogContext);
-              },
-              child: const Text('Cancel'),
-            ),
-            ElevatedButton(
-              onPressed: () async {
-                if (titleController.text.isEmpty) {
-                  ScaffoldMessenger.of(dialogContext).showSnackBar(
-                    const SnackBar(
-                      content: Text('Please enter an appointment title'),
-                    ),
-                  );
-                  return;
-                }
-
-                final newAppointment = Appointment(
-                  date: selectedDate,
-                  title: titleController.text,
-                  description: descriptionController.text,
-                  isPast: selectedDate.isBefore(DateTime.now()),
-                );
-
-                final dialogCtx = dialogContext;
-                final success =
-                    await DiaryService.saveAppointment(context, newAppointment);
-                if (success) {
-                  if (mounted) {
-                    parentSetState(() {
-                      appointments.add(newAppointment);
-                    });
-                  }
-                  if (dialogCtx.mounted) {
-                    Navigator.pop(dialogCtx);
-                  }
-                }
-              },
-              child: const Text('Add'),
-            ),
-          ],
         );
       },
     );
@@ -500,127 +190,20 @@ class _AppointmentCardState extends State<AppointmentCard> {
           if (_isLoading)
             const Center(child: CircularProgressIndicator())
           else if (appointments.isNotEmpty) ...[
-            Text(
-              subtitle,
-              style: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.w500,
-              ),
+            AppointmentInfoDisplay(
+              appointment: appointments.first,
+              subtitle: subtitle,
             ),
             const SizedBox(height: 16),
-            // Date.
-
-            _buildInfoRow(
-              'Date:',
-              'Monday, ${DateFormat('d MMMM').format(appointments.first.date)}',
+            TransportInfoWidget(
+              useClinicBus: useClinicBus,
+              needsTransport: needsTransport,
+              transportPhone: transportPhone,
+              transportNote: transportNote,
             ),
-            const SizedBox(height: 8),
-            // Time.
-
-            _buildInfoRow(
-              'Time:',
-              DateFormat('h:mm a').format(appointments.first.date),
-            ),
-            const SizedBox(height: 8),
-            // Description.
-
-            _buildInfoRow('Description:', appointments.first.description),
-            const SizedBox(height: 16),
-            // Transport.
-
-            if (useClinicBus)
-              Row(
-                children: [
-                  const Icon(Icons.directions_bus, color: Colors.green),
-                  const SizedBox(width: 8),
-                  const Text(
-                    'Transport:',
-                    style: TextStyle(fontWeight: FontWeight.bold),
-                  ),
-                  const Icon(Icons.check, color: Colors.green),
-                ],
-              ),
-            if (needsTransport) ...[
-              const SizedBox(height: 16),
-              const Text(
-                'Need help with transport?',
-                style: TextStyle(fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 8),
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  const Icon(Icons.phone, size: 20),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text.rich(
-                      TextSpan(
-                        children: [
-                          TextSpan(
-                            text: 'Call $transportPhone ',
-                            style: const TextStyle(fontWeight: FontWeight.bold),
-                          ),
-                          TextSpan(
-                            text: transportNote,
-                            style: const TextStyle(fontStyle: FontStyle.italic),
-                          ),
-                          const TextSpan(
-                            text: ' to change or request transport.',
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                  SizedBox(
-                    width: 32,
-                    height: 32,
-                    child: MarkdownTooltip(
-                      message: _isPlaying
-                          ? '**Stop** audio'
-                          : '**Play** audio explanation',
-                      child: IconButton(
-                        icon: Icon(
-                          _isPlaying ? Icons.stop : Icons.volume_up,
-                          color: _isPlaying ? Colors.red : Colors.blue,
-                          size: 20,
-                        ),
-                        onPressed: _toggleAudio,
-                        padding: EdgeInsets.zero,
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ],
           ],
         ],
       ),
-    );
-  }
-
-  /// Helper method to build consistent information rows.
-  ///
-  /// Creates a row with a label and value, maintaining consistent styling
-  /// and layout across the card.
-
-  Widget _buildInfoRow(String label, String value) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SizedBox(
-          width: 90,
-          child: Text(
-            label,
-            style: const TextStyle(
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ),
-        Expanded(
-          child: Text(value),
-        ),
-      ],
     );
   }
 }
