+++ lib/features/home/service/components/dialogs/add_appointment_dialog.dart
@@ -1,0 +1,184 @@
+/// Dialog for adding new appointments.
+///
+/// Copyright (C) 2025, Software Innovation Institute, ANU
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://www.gnu.org/licenses/gpl-3.0.en.html
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://www.gnu.org/licenses/>.
+///
+/// Authors: Zheyuan Xu, Tony Chen
+
+library;
+
+import 'package:flutter/material.dart';
+
+import 'package:healthpod/features/diary/models/appointment.dart';
+import 'package:healthpod/features/diary/service.dart';
+
+/// Dialog for adding new appointments with form fields for all appointment
+/// details.
+
+class AddAppointmentDialog extends StatefulWidget {
+  const AddAppointmentDialog({
+    super.key,
+    required this.onAppointmentAdded,
+  });
+
+  final Function(Appointment appointment) onAppointmentAdded;
+
+  @override
+  State<AddAppointmentDialog> createState() => _AddAppointmentDialogState();
+}
+
+class _AddAppointmentDialogState extends State<AddAppointmentDialog> {
+  final TextEditingController titleController = TextEditingController();
+  final TextEditingController descriptionController = TextEditingController();
+  DateTime selectedDate = DateTime.now();
+  TimeOfDay selectedTime = TimeOfDay.now();
+
+  @override
+  void dispose() {
+    titleController.dispose();
+    descriptionController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text('Add New Appointment'),
+      content: SingleChildScrollView(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            TextField(
+              controller: titleController,
+              decoration: const InputDecoration(
+                labelText: 'Appointment Title',
+                border: OutlineInputBorder(),
+              ),
+            ),
+            const SizedBox(height: 16),
+            TextField(
+              controller: descriptionController,
+              decoration: const InputDecoration(
+                labelText: 'Description',
+                border: OutlineInputBorder(),
+              ),
+            ),
+            const SizedBox(height: 16),
+            Row(
+              children: [
+                Expanded(
+                  child: OutlinedButton.icon(
+                    onPressed: _selectDate,
+                    icon: const Icon(Icons.calendar_today),
+                    label: const Text('Set Date'),
+                  ),
+                ),
+                const SizedBox(width: 8),
+                Expanded(
+                  child: OutlinedButton.icon(
+                    onPressed: _selectTime,
+                    icon: const Icon(Icons.access_time),
+                    label: const Text('Set Time'),
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+      ),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.pop(context),
+          child: const Text('Cancel'),
+        ),
+        ElevatedButton(
+          onPressed: _addAppointment,
+          child: const Text('Add'),
+        ),
+      ],
+    );
+  }
+
+  /// Shows date picker for appointment date selection.
+
+  Future<void> _selectDate() async {
+    final DateTime? picked = await showDatePicker(
+      context: context,
+      initialDate: selectedDate,
+      firstDate: DateTime.now(),
+      lastDate: DateTime(2100),
+    );
+    if (picked != null) {
+      setState(() {
+        selectedDate = DateTime(
+          picked.year,
+          picked.month,
+          picked.day,
+          selectedTime.hour,
+          selectedTime.minute,
+        );
+      });
+    }
+  }
+
+  /// Shows time picker for appointment time selection.
+
+  Future<void> _selectTime() async {
+    final TimeOfDay? picked = await showTimePicker(
+      context: context,
+      initialTime: selectedTime,
+    );
+    if (picked != null) {
+      setState(() {
+        selectedTime = picked;
+        selectedDate = DateTime(
+          selectedDate.year,
+          selectedDate.month,
+          selectedDate.day,
+          selectedTime.hour,
+          selectedTime.minute,
+        );
+      });
+    }
+  }
+
+  /// Creates and saves the new appointment.
+
+  Future<void> _addAppointment() async {
+    if (titleController.text.isEmpty) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Please enter an appointment title')),
+      );
+      return;
+    }
+
+    final newAppointment = Appointment(
+      date: selectedDate,
+      title: titleController.text,
+      description: descriptionController.text,
+      isPast: selectedDate.isBefore(DateTime.now()),
+    );
+
+    final success = await DiaryService.saveAppointment(context, newAppointment);
+    if (success && mounted) {
+      widget.onAppointmentAdded(newAppointment);
+      Navigator.pop(context);
+    }
+  }
+}
