+++ lib/features/table/bp_editor/page.dart
@@ -21,19 +21,17 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <https://www.gnu.org/licenses/>.
 ///
-/// Authors: Ashley Tang
+/// Authors: Ashley Tang, Tony Chen
 
 library;
 
 import 'package:flutter/material.dart';
 
-import 'package:intl/intl.dart';
-
 import 'package:healthpod/features/bp/obs/model.dart';
 import 'package:healthpod/features/bp/obs/service.dart';
-import 'package:healthpod/features/bp/obs/widgets/display_row.dart';
-import 'package:healthpod/features/bp/obs/widgets/editing_row.dart';
 import 'package:healthpod/features/table/bp_editor/state.dart';
+import 'package:healthpod/features/table/bp_editor/widgets/desktop_layout.dart';
+import 'package:healthpod/features/table/bp_editor/widgets/mobile_layout.dart';
 
 /// The main editor page for blood pressure observations.
 
@@ -121,409 +119,94 @@ class _BPEditorPageState extends State<BPEditorPage> {
     });
   }
 
+  /// Handles saving an observation.
+
+  Future<void> _handleSave(int index) async {
+    await editorState.saveObservation(
+      context,
+      editorService,
+      index,
+    );
+    _loadData();
+  }
+
+  /// Handles deleting an observation.
+
+  Future<void> _handleDelete(BPObservation obs, int index) async {
+    try {
+      await editorState.deleteObservation(
+        context,
+        editorService,
+        obs,
+      );
+
+      if (context.mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(
+            content: Text(
+              'Blood pressure reading deleted successfully.',
+            ),
+            backgroundColor: Colors.green,
+          ),
+        );
+      }
+    } catch (e) {
+      if (context.mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text('Error deleting reading: ${e.toString()}'),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+    } finally {
+      // Always reload data to reflect current state.
+
+      _loadData();
+    }
+  }
+
+  /// Handles timestamp changes during editing.
+
+  void _handleTimestampChanged(DateTime newTimestamp) {
+    setState(() {
+      editorState.currentEdit = editorState.currentEdit?.copyWith(
+        timestamp: newTimestamp,
+      );
+    });
+  }
+
   /// Builds the desktop layout for the blood pressure editor.
-  ///
-  /// Uses a DataTable with responsive columns that adapt based on screen width.
-  /// The table shows a minimum set of columns (timestamp, systolic, diastolic) and
-  /// progressively reveals more columns (heart rate, notes) as screen width increases.
-  ///
-  /// @param context The build context.
-  /// @param width The current screen width.
-  /// @returns A Widget containing the desktop layout.
 
   Widget _buildDesktopLayout(BuildContext context, double width) {
-    final columns = _getColumns(width);
-
-    return SingleChildScrollView(
-      scrollDirection: Axis.horizontal,
-      child: SingleChildScrollView(
-        child: DataTable(
-          columns: columns,
-          rows: List<DataRow>.generate(
-            editorState.observations.length,
-            (index) {
-              final obs = editorState.observations[index];
-
-              if (editorState.editingIndex == index) {
-                return buildEditingRow(
-                  context: context,
-                  width: width,
-                  editorState: editorState,
-                  editorService: editorService,
-                  observation: obs,
-                  index: index,
-                  onCancel: _handleCancelEdit,
-                  onSave: () async {
-                    await editorState.saveObservation(
-                      context,
-                      editorService,
-                      index,
-                    );
-                    _loadData();
-                  },
-                  onTimestampChanged: (DateTime newTimestamp) {
-                    setState(() {
-                      editorState.currentEdit =
-                          editorState.currentEdit?.copyWith(
-                        timestamp: newTimestamp,
-                      );
-                    });
-                  },
-                );
-              }
-
-              return buildDisplayRow(
-                context: context,
-                width: width,
-                observation: obs,
-                index: index,
-                onEdit: () => setState(() {
-                  editorState.enterEditMode(index);
-                }),
-                onDelete: () async {
-                  try {
-                    await editorState.deleteObservation(
-                      context,
-                      editorService,
-                      obs,
-                    );
-
-                    if (context.mounted) {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        const SnackBar(
-                          content: Text(
-                            'Blood pressure reading deleted successfully.',
-                          ),
-                          backgroundColor: Colors.green,
-                        ),
-                      );
-                    }
-                  } catch (e) {
-                    if (context.mounted) {
-                      ScaffoldMessenger.of(context).showSnackBar(
-                        SnackBar(
-                          content:
-                              Text('Error deleting reading: ${e.toString()}'),
-                          backgroundColor: Colors.red,
-                        ),
-                      );
-                    }
-                  } finally {
-                    // Always reload data to reflect current state.
-
-                    _loadData();
-                  }
-                },
-              );
-            },
-          ),
-        ),
-      ),
+    return BPEditorDesktopLayout(
+      editorState: editorState,
+      editorService: editorService,
+      width: width,
+      onEdit: (index) => () => setState(() {
+            editorState.enterEditMode(index);
+          }),
+      onDelete: _handleDelete,
+      onCancelEdit: _handleCancelEdit,
+      onSave: _handleSave,
+      onTimestampChanged: _handleTimestampChanged,
     );
   }
 
   /// Builds the mobile layout for the blood pressure editor.
-  ///
-  /// Creates a scrollable list of cards, each representing a blood pressure observation.
-  /// Cards can be expanded to show additional details and include edit/delete actions.
-  ///
-  /// @param context The build context.
-  /// @returns A Widget containing the mobile layout.
 
   Widget _buildMobileLayout(BuildContext context) {
-    return ListView.builder(
-      itemCount: editorState.observations.length,
-      itemBuilder: (context, index) {
-        final obs = editorState.observations[index];
-        final isEditing = editorState.editingIndex == index;
-
-        if (isEditing) {
-          return _buildMobileEditCard(obs, index);
-        }
-
-        return Card(
-          margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-          child: ExpansionTile(
-            title: Text(
-              DateFormat('yyyy-MM-dd HH:mm').format(obs.timestamp),
-              style: Theme.of(context).textTheme.titleMedium,
-            ),
-            subtitle: Text(
-              'BP: ${obs.systolic.round()}/${obs.diastolic.round()} mmHg',
-              style: Theme.of(context).textTheme.bodyMedium,
-            ),
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    _buildInfoRow('Heart Rate', '${obs.heartRate.round()} BPM'),
-                    if (obs.notes.isNotEmpty) ...[
-                      const SizedBox(height: 8),
-                      _buildInfoRow('Notes', obs.notes),
-                    ],
-                    const SizedBox(height: 16),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.end,
-                      children: [
-                        IconButton(
-                          icon: const Icon(Icons.edit),
-                          onPressed: () => setState(() {
-                            editorState.enterEditMode(index);
-                          }),
-                        ),
-                        IconButton(
-                          icon: const Icon(Icons.delete),
-                          onPressed: () async {
-                            try {
-                              await editorState.deleteObservation(
-                                context,
-                                editorService,
-                                obs,
-                              );
-
-                              if (context.mounted) {
-                                ScaffoldMessenger.of(context).showSnackBar(
-                                  const SnackBar(
-                                    content: Text(
-                                      'Blood pressure reading deleted successfully.',
-                                    ),
-                                    backgroundColor: Colors.green,
-                                  ),
-                                );
-                              }
-                            } catch (e) {
-                              if (context.mounted) {
-                                ScaffoldMessenger.of(context).showSnackBar(
-                                  SnackBar(
-                                    content: Text(
-                                      'Error deleting reading: ${e.toString()}',
-                                    ),
-                                    backgroundColor: Colors.red,
-                                  ),
-                                );
-                              }
-                            } finally {
-                              // Always reload data to reflect current state.
-
-                              _loadData();
-                            }
-                          },
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
+    return BPEditorMobileLayout(
+      editorState: editorState,
+      editorService: editorService,
+      onEdit: (index) => () => setState(() {
+            editorState.enterEditMode(index);
+          }),
+      onDelete: _handleDelete,
+      onCancelEdit: _handleCancelEdit,
+      onSave: _handleSave,
+      onTimestampChanged: _handleTimestampChanged,
     );
-  }
-
-  /// Builds a mobile-optimised numeric input field.
-
-  Widget _buildMobileNumericField({
-    required TextEditingController? controller,
-    required String label,
-    required String suffix,
-  }) {
-    return TextField(
-      controller: controller,
-      keyboardType: TextInputType.number,
-      decoration: InputDecoration(
-        labelText: label,
-        suffixText: suffix,
-      ),
-    );
-  }
-
-  /// Builds an editing card for the mobile layout.
-  ///
-  /// Creates a form-style card with fields for editing all observation properties.
-  /// Optimized for touch interaction and mobile screen sizes.
-  ///
-  /// @param obs The blood pressure observation being edited.
-  /// @param index The index of the observation in the list.
-  /// @returns A Widget containing the mobile editing form.
-
-  Widget _buildMobileEditCard(BPObservation obs, int index) {
-    return Card(
-      margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              'Edit Reading',
-              style: Theme.of(context).textTheme.titleLarge,
-            ),
-            const SizedBox(height: 16),
-            InkWell(
-              onTap: () async {
-                final date = await showDatePicker(
-                  context: context,
-                  initialDate: obs.timestamp,
-                  firstDate: DateTime(2000),
-                  lastDate: DateTime.now(),
-                );
-                if (!context.mounted) return;
-
-                if (date != null) {
-                  if (!mounted) return;
-                  final time = await showTimePicker(
-                    context: context,
-                    initialTime: TimeOfDay.fromDateTime(obs.timestamp),
-                  );
-                  if (time != null && context.mounted) {
-                    final newTimestamp = DateTime(
-                      date.year,
-                      date.month,
-                      date.day,
-                      time.hour,
-                      time.minute,
-                    );
-                    setState(() {
-                      editorState.currentEdit =
-                          editorState.currentEdit?.copyWith(
-                        timestamp: newTimestamp,
-                      );
-                    });
-                  }
-                }
-              },
-              child: _buildInfoRow(
-                'Date/Time',
-                DateFormat('yyyy-MM-dd HH:mm').format(
-                  editorState.currentEdit?.timestamp ?? obs.timestamp,
-                ),
-                isEditable: true,
-              ),
-            ),
-            const SizedBox(height: 16),
-            _buildMobileNumericField(
-              controller: editorState.systolicController,
-              label: 'Systolic',
-              suffix: 'mmHg',
-            ),
-            const SizedBox(height: 8),
-            _buildMobileNumericField(
-              controller: editorState.diastolicController,
-              label: 'Diastolic',
-              suffix: 'mmHg',
-            ),
-            const SizedBox(height: 8),
-            _buildMobileNumericField(
-              controller: editorState.heartRateController,
-              label: 'Heart Rate',
-              suffix: 'BPM',
-            ),
-            const SizedBox(height: 16),
-            TextField(
-              controller: editorState.notesController,
-              maxLines: 3,
-              decoration: const InputDecoration(
-                labelText: 'Notes',
-                alignLabelWithHint: true,
-              ),
-            ),
-            const SizedBox(height: 24),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: [
-                TextButton(
-                  onPressed: _handleCancelEdit,
-                  child: const Text('Cancel'),
-                ),
-                const SizedBox(width: 16),
-                ElevatedButton(
-                  onPressed: () async {
-                    await editorState.saveObservation(
-                      context,
-                      editorService,
-                      index,
-                    );
-                    _loadData();
-                  },
-                  child: const Text('Save'),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  /// Builds a consistent row layout for displaying information.
-  ///
-  /// Creates a two-column layout with a label and value, optionally styling the value
-  /// to indicate it is editable.
-  ///
-  /// @param label The label text to display.
-  /// @param value The value text to display.
-  /// @param isEditable Whether to style the value as an editable field.
-  /// @returns A Widget containing the formatted information row.
-
-  Widget _buildInfoRow(String label, String value, {bool isEditable = false}) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        SizedBox(
-          width: 100,
-          child: Text(
-            label,
-            style: TextStyle(
-              fontWeight: FontWeight.bold,
-              color: Theme.of(context).colorScheme.primary,
-            ),
-          ),
-        ),
-        Expanded(
-          child: Text(
-            value,
-            style: TextStyle(
-              decoration: isEditable ? TextDecoration.underline : null,
-              color: isEditable ? Colors.blue : null,
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-
-  /// Gets the list of columns to display in the DataTable based on screen width.
-  ///
-  /// Implements responsive column visibility:
-  /// - Base columns (Timestamp, Systolic, Diastolic) always visible
-  /// - Heart Rate visible when width > 600
-  /// - Notes visible when width > 800
-  /// - Actions column always visible
-  ///
-  /// @param width The current screen width.
-  /// @returns A list of DataColumn objects.
-
-  List<DataColumn> _getColumns(double width) {
-    final List<DataColumn> columns = [
-      const DataColumn(label: Text('Timestamp')),
-      const DataColumn(label: Text('Systolic')),
-      const DataColumn(label: Text('Diastolic')),
-    ];
-
-    if (width > 600) {
-      columns.add(const DataColumn(label: Text('Heart Rate')));
-    }
-
-    if (width > 800) {
-      columns.add(const DataColumn(label: Text('Notes')));
-    }
-
-    columns.add(const DataColumn(label: Text('Actions')));
-
-    return columns;
   }
 
   @override
