+++ macos/Runner/DebugProfile.entitlements
@@ -10,10 +10,6 @@
 	<true/>
 	<key>com.apple.security.network.client</key>
     <true/>
-	<key>keychain-access-groups</key>
-    <array/>
-	<key>com.apple.security.keychain</key>
-    <true/>
 	<key>com.apple.security.files.user-selected.read-write</key>
     <true/>
 </dict>
